alphabetical order + time and update libft

+++ Libft/src/ft_putnbr_fd.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long n, int fd)
 {
-	unsigned int	x;
+	unsigned long	x;
 	char			c;
 
 	if (n < 0)
